Move npm helpers to lib/util
